refactor : 인터페이스 패키지변경
