+++ src/main/java/nextstep/ladder/domain/LadderWidth.java
@@ -2,6 +2,8 @@ package nextstep.ladder.domain;
 
 import java.util.ArrayList;
 import java.util.List;
+import nextstep.ladder.domain.validator.MinSizeValidator;
+import nextstep.ladder.interfaces.Position;
 
 public class LadderWidth {
     private static final MinSizeValidator minSizeValidator = new MinSizeValidator();
