+++ src/main/java/nextstep/ladder/domain/LineRange.java
@@ -1,6 +1,7 @@
 package nextstep.ladder.domain;
 
 import java.util.Objects;
+import nextstep.ladder.interfaces.Position;
 
 public class LineRange {
     private final int left;
