+++ src/main/java/nextstep/ladder/interfaces/Lines.java
@@ -1,4 +1,4 @@
-package nextstep.ladder.domain;
+package nextstep.ladder.interfaces;
 
 import java.util.List;
