+++ src/test/java/nextstep/ladder/domain/LineRangeTest.java
@@ -5,6 +5,7 @@ import static org.assertj.core.api.Assertions.assertThatIllegalArgumentException
 
 import java.util.List;
 import java.util.stream.Stream;
+import nextstep.ladder.interfaces.Position;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.Arguments;
